requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pillow==10.0.0
numpy==1.26.4
torch==2.0.0
torchvision==0.15.1
